+++ src/main/kotlin/de/achimonline/ansible_lint/annotator/actions/AnsibleLintAnnotatorClipboardAction.kt
@@ -9,7 +9,8 @@ import com.intellij.psi.PsiFile
 import com.intellij.util.ui.TextTransferable
 import de.achimonline.ansible_lint.bundle.AnsibleLintBundle.message
 
-class AnsibleLintAnnotatorClipboardAction(private val ruleId: String) : AnsibleLintAnnotatorAction(message("action.copy-rule-id-to-clipboard")) {
+class AnsibleLintAnnotatorClipboardAction(private val ruleId: String) :
+    AnsibleLintAnnotatorAction(message("action.copy-rule-id-to-clipboard")) {
     override fun invoke(project: Project, editor: Editor?, file: PsiFile?) {
         ClipboardSynchronizer
             .getInstance()