+++ src/test/kotlin/de/achimonline/ansible_lint/settings/AnsibleLintConfigurableTest.kt
@@ -8,8 +8,10 @@ import org.junit.Assert.*
 class AnsibleLintConfigurableTest {
     @Test
     fun executeTest_icons() {
-        // make sure the icons (which are being used in an HTML-string) are still available
+        // make sure the icons (which are being used in an HTML-strings) are still available
+        assertNotNull(AllIcons.General.Information)
         assertNotNull(AllIcons.General.InspectionsOK)
-        assertNotNull(AllIcons.General.InspectionsError)
+        assertNotNull(AllIcons.General.NotificationWarning)
+        assertNotNull(AllIcons.General.NotificationError)
     }
 }